refactor(api): neutralize messaging service filenames
Rename four messaging service modules to domain-neutral filenames and
update every referencing module. All exports in the four files were
already neutral, so the moves are pure 100% renames with no body changes.

- services/zero-telegram-link.service.ts -> services/telegram-link.service.ts
- services/zero-teams-dispatch.service.ts -> services/teams-dispatch.service.ts
- services/zero-feishu-welcome.service.ts -> services/feishu-welcome.service.ts
- services/zero-feishu-connect.service.ts -> services/feishu-connect.service.ts

Seven import specifiers move to the new paths. The agent compose
consolidation preflight manifest pins scanned repository paths, so its
six entries for three of the renamed files move with them; the entries
are otherwise unchanged and stay in collector sort order.
