+++ turbo/apps/api/src/signals/services/cron-cleanup-sandboxes.service.ts
@@ -27,7 +27,7 @@ import { drainStaleChatThreadQueues$ } from "./chat-thread-queue-drain.service";
 import type { QueueMarkerRevokeNotification } from "./chat-queue-marker.service";
 import { drainStaleCanonicalSlackIngress$ } from "./canonical-slack-ingress-processor.service";
 import { drainStaleCanonicalFeishuIngress$ } from "./canonical-feishu-ingress-processor.service";
-import { retryPendingFeishuConnectWelcomes$ } from "./zero-feishu-welcome.service";
+import { retryPendingFeishuConnectWelcomes$ } from "./feishu-welcome.service";
 import {
   cleanupThreadlessRuns$,
   type ThreadlessRunCleanupResult,
