+++ turbo/packages/db/scripts/agent-compose-consolidation-preflight-manifest.ts
@@ -181,7 +181,9 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/cron-cleanup-sandboxes.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/cron-compact-chat-thread-snapshots.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/cron-project-chat-event-search.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
+    "turbo/apps/api/src/signals/services/feishu-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/feishu-dispatch.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
+    "turbo/apps/api/src/signals/services/feishu-welcome.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/github-agent-reply-footer.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/github-agent-reply-footer.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/github-oauth.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
@@ -200,6 +202,7 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/shared-thread.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/slack-message-context.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/slack-webhooks.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
+    "turbo/apps/api/src/signals/services/teams-dispatch.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/telegram-data.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/telegram-footer.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/telegram-footer.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
@@ -216,8 +219,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-browser.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-compose-data.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
-    "turbo/apps/api/src/signals/services/zero-feishu-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-feishu-welcome.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
@@ -227,7 +228,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-slack-data.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-teams-dispatch.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/packages/db/src/schema/agent-draft.ts|./zero-agent|zeroAgents:zeroAgents",
@@ -302,7 +302,9 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/cron-steer-run-time-budget.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/cron-sync-skills.service.ts|headVersionId",
     "turbo/apps/api/src/signals/services/feishu-chat-ingress.service.ts|agentComposeId",
+    "turbo/apps/api/src/signals/services/feishu-connect.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/feishu-dispatch.service.ts|zeroAgents",
+    "turbo/apps/api/src/signals/services/feishu-welcome.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/github-agent-reply-footer.service.ts|agentComposes,zeroAgents",
     "turbo/apps/api/src/signals/services/github-oauth.service.ts|agentComposes",
     "turbo/apps/api/src/signals/services/github-queued-launch-context.service.ts|agentComposeId",
@@ -330,6 +332,7 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/storage-volume-publication.service.ts|headVersionId",
     "turbo/apps/api/src/signals/services/storage-write.service.ts|headVersionId",
     "turbo/apps/api/src/signals/services/teams-chat-ingress.service.ts|agentComposeId",
+    "turbo/apps/api/src/signals/services/teams-dispatch.service.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/telegram-chat-ingress.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/telegram-data.service.ts|agentComposeVersions,agentComposes,headVersionId",
     "turbo/apps/api/src/signals/services/telegram-footer.service.ts|agentComposeId,agentComposes,zeroAgents",
@@ -347,14 +350,11 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-chat-title.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/zero-compose-data.service.ts|agentComposeId,agentComposeVersions,agentComposes",
-    "turbo/apps/api/src/signals/services/zero-feishu-connect.service.ts|zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-feishu-welcome.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|agentComposeId,agentComposeVersionId,agentComposeVersions,agentComposes,headVersionId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs.service.ts|agentComposeId,agentComposeVersionId,agentComposes,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|agentComposes,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-slack-data.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|agentComposeVersions,agentComposes,headVersionId,zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-teams-dispatch.service.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|agentComposeId,agentComposes,zeroAgents",
     "turbo/apps/platform/src/views/zero-page/zero-telegram-settings-page.tsx|headVersionId",
     "turbo/packages/api-contracts/src/contracts/composes.ts|headVersionId",
